backend/requirements.txt
fastapi
uvicorn
python-multipart
Pillow
opencv-python
numpy
imagehash
invisible-watermark
moviepy
algosdk
